// src/Components/model/CouponDetails.ts
export default class CouponDetails {
    id: number = 0;
    companyID: number = 0;
    categoryID: number = 0;
    title: string = "";
    discount: string="";
    description: string = "";
    startDate: Date = new Date();
    endDate: Date = new Date();
    amount: number = 0;
    price: number = 0;
    image: string = "";
}